backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlmodel==0.0.24
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
openpyxl==3.1.5
